{{ cookiecutter.project_name }}/arc/requirements.txt
alembic==1.12.0
allure-behave==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==2.11.1
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.11.1
behave==1.2.6
behave-pytest==0.1.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.5.1
junit2html==30.1.3
lxml==4.9.3
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.0.9
outcome==1.3.0
packaging==23.2
pandas==2.0.1
parse==1.19.1
parse-type==0.6.2
Pillow==10.0.1
pluggy==1.3.0
psutil==5.9.7
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.16.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==2.9.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.1
scikit-learn==1.3.1
scipy==1.11.4
screeninfo==0.8
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
teamcity-messages==1.31
threadpoolctl==3.2.0
trio==0.22.2
trio-websocket==0.11.1
typer==0.7.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.0
